Add CUDA lib to Makefile

+++ src/seqio.cpp
@@ -136,12 +136,12 @@ std::vector<char> SeqBuf::as_square_array(const size_t n_threads) const {
 
     std::vector<char> read_array(_max_length * _full_index.size());
     #pragma omp parallel for simd schedule(static) num_threads(n_threads)
-    for (long int read_idx = 0; read_idx < _full_index.size(); read_idx++) {
+    for (size_t read_idx = 0; read_idx < _full_index.size(); read_idx++) {
         std::string seq = sequence[_full_index[read_idx]];
-        for (int base_idx = 0; base_idx < seq.size(); base_idx++) {
+        for (size_t base_idx = 0; base_idx < seq.size(); base_idx++) {
             read_array[read_idx + base_idx * _max_length] = seq[base_idx];
         }
-        for (int base_idx = seq.size(); base_idx < _max_length; base_idx++) {
+        for (size_t base_idx = seq.size(); base_idx < _max_length; base_idx++) {
             read_array[read_idx + base_idx * _max_length] = 'N';
         }
     }
